changed according to PR comments
