+++ example/styles/index.ts
@@ -1,0 +1,140 @@
+import {STYLES} from '@likeminds.community/chat-rn-core';
+
+export const setStyles = () => {
+  // themeStyling
+  {
+    /*
+  const themeStyles = {
+    hue: 10,
+    fontColor: 'black',
+    primaryColor: 'green',
+    secondaryColor: 'green',
+    lightBackgroundColor: '#d7f7ed',
+  };
+  */
+  }
+
+  // styling for reactionList
+  {
+    /*
+  const reactionListStyles = {
+    reactionSize: 0,
+    reactionLeftItemStroke: 'pink',
+    reactionRightItemStroke: 'yellow',
+    reactionItemBorderRadius: 5,
+    gap: 5,
+  };
+  */
+  }
+
+  // styling for chatBubble
+  {
+    /*
+  const chatBubbleStyles = {
+    borderRadius: 5,
+    sentMessageBackgroundColor: 'yellow',
+    receivedMessageBackgroundColor: 'pink',
+    selectedBackgroundColor: 'grey',
+    selectedMessageBackgroundColor: 'purple',
+    textStyles: {
+      fontSize: 10,
+      fontStyle: 'italic',
+      fontFamily: 'SofiaPro-SemiBold',
+    },
+    linkTextColor: 'red',
+    taggingTextColor: 'yellow',
+    stateMessagesBackgroundColor: 'pink',
+    stateMessagesTextStyles: {
+      fontSize: 10,
+      fontStyle: 'italic',
+      fontFamily: 'SofiaPro-SemiBold',
+    },
+  };
+  */
+  }
+
+  // styling for inputBox
+  {
+    /*
+  const inputBoxStyles = {
+    placeholderTextColor: '#aaa',
+    selectionColor: '#aaa',
+    plainTextStyle: {color: 'black'},
+    partsTextStyle: {
+      color: '#007AFF',
+    },
+    sendIconStyles: {
+      width: 22,
+      height: 22,
+      resizeMode: 'contain',
+      marginLeft: 5,
+    },
+    attachmentIconStyles: {
+      width: 22,
+      height: 22,
+      resizeMode: 'contain',
+    },
+    micIconStyles: {
+      width: 22,
+      height: 22,
+      resizeMode: 'contain',
+    },
+    cameraIconStyles: {
+      width: 22,
+      height: 22,
+      resizeMode: 'contain',
+    },
+    galleryIconStyles: {
+      width: 22,
+      height: 22,
+      resizeMode: 'contain',
+    },
+    documentIconStyles: {
+      width: 22,
+      height: 22,
+      resizeMode: 'contain',
+    },
+    pollIconStyles: {
+      width: 22,
+      height: 22,
+      resizeMode: 'contain',
+    },
+  };
+  */
+  }
+
+  // styling for chatroom topic
+  {
+    /* 
+      const chatroomTopicStyles = {
+    topicHeader: {
+      color: 'blue',
+    },
+    topicDescription: {
+      color: 'yellow',
+    },
+    topicPlaceholder: 'Topic set by your CM',
+  };
+    */
+  }
+
+  // if (chatBubbleStyles) {
+  //   STYLES.setChatBubbleStyle(chatBubbleStyles);
+  // }
+
+  // if (themeStyles) {
+  //   STYLES.setTheme(themeStyles);
+  // }
+
+  // if (reactionListStyles) {
+  //   STYLES.setReactionListStyle(reactionListStyles);
+  // }
+
+  // if (inputBoxStyles) {
+  //   STYLES.setInputBoxStyle(inputBoxStyles);
+  // }
+
+  // if (chatroomTopicStyles) {
+  //   STYLES.setChatroomTopicStyle(chatroomTopicStyles);
+  // }
+};
